Started the README.md for my profile
Added a brief introduction including some important personal information.
Added a section for links to social media profiles and professional accounts.
Added a section containing images of the many languages and tools that I have used to complete projects over the years.
Decided to download the SVG images and upload them in a separate folder for hosting and longevity purposes.